Update to current mruby VM
It looks like everything is fairly stable, so this may resolve some issues
present in long lived ardour sessions. The underlying assumption is that some
pre-existing issues related to memory instability or GC bugs, though they may
be unrelated.

It does appear that there may be a regression in the accuracy of line numbering
in backtraces and it is suspected that the earlier profiling techniques might
not apply due to the VM instruction changed over the past 2 years.