Update MRuby version to an ~3.0.0 version
The update was motivated by the fact that optional keyword arguments are
handled slightly under Ruby 3.0.0 (not the embedded one, but the system
one used to bootstrap). The MRuby source could be patched up easy
enough, but keeping deps in line is good.

Thankfully this was a lot smoother than older MRuby updates were.
instance_eval appears to work fine (the GUI is all sorts of broken when
that was buggy in the past), and for the most part things just worked.
For some reason the mruby-file-stat gem is acting up, but that was easy
enough to hack into the makefile.

The biggest change was in integer division. Previously it cast the
result to floating point, but now it keeps the result as an integer
(seems to match the behavior in non-MRuby).
There were a bunch of spots which required some minor fixing, but all
the views seem to look like they render correctly.

The more typical breakage is backtraces aren't working again. I don't
see anything open on the issue tracker, so perhaps this is a problem on
our side, but it's been a problem on their side a number of times.