Import pugl directly into root repo
We're at the stage where we're bypassing waf, we don't use cairo, etc.
So once MacOS is handled, then things can be boiled down to the core
pugl.h, pugl_win.cpp, pugl_x11.c, and pugl_osx.m

For those without context: the waf included relies on python2.7 code,
which is outstandingly out-of-date at this point and was old even when
this version of pugl was first integrated. There are additional upstream
changes, but integrating them is non-trivial and doesn't necessarily
present users with advantages.