mruby-widget-lib: api: Add API for updating Zest UI
DAWs like REAPER support saving VST's state as so-called "user presets"
 (or directly saving it into project file). And in most VST plugins,
when you load a "user preset", the state (aka. configurations) of
that plugin is restored, and the UI is also updated according to the state.

But for a long time, Zyn-Fusion doesn't support this feature. This will
be troublesome for DAW users. To solve this, I added a method
zest_forget_all_state(). It will send OSC message "damage" to inform
ZynAddSubFX that the UI is outdated, and an update is needed.

Invoke this method in DPF plugin's DISTRHO::UI::stateChanged() so that
every time VST host loads presets can Zest UI be updated
correspondingly.

Thanks @fundamental for instructions.

Related issue: #35