Advance windows build
Right now it looks like the windows build might be working correctly.
The string backtrace patch is required for the 'superhack' which amounts
to taking exception information normally streamed to stderr and leaving
it in a global variable. That makes it possible to present error
messages to users.

Additional changes may be needed in the makefile including perhaps some
inlining of the mingw-* tool names. I'd prefer not to do so since I can
imagine some cases where they might not be in the path or other system
quirks, but if it's needed I'm ready to do so to get this next release
out.