Restructure Instrument settings and Portamento
Restructure Instrument Settings and Portamento panels, moving the
Portamento toggle button from Instrument Settings to
Portamento, in order to make space for the new vlimit NumEntry
widget.

Instead of having the label for vlimit in ultra small font
below the actual NumEntry widget, put it as a text box to the
left of it. By shorting the names of some of the key modes, we
get a nice last line in the Instrument Settings box.

Restructure Portamento box to get all NumEntry and Toggle
widgets the same size, which looks neater.