Fix windows build
As clock_gettime isn't available through mingw builds (based on the
linker errors), but mingw makes it look like it's available I'm
resorting to messing with POSIX defines.
This seems to be enough to convince mruby-time that the function isn't
usable which shifts it over to gettimeofday() which should be fine.