ZynAmpEnv: Put all ADSR parameters in a row (#311)
As per discussion in issue #311, put all envelope parameters
in a row for the ADSR (amplitude) envelopes, thus making the
envelope layout more conventional.

Also adjust the position of the toggle buttons, putting FRCR and
REPEAT over each other, like in the Filter Envelope, and LIN/LOG
beside it, to avoid the three toggle buttons being rendered
as thin stripes.