Zynbank: Clear bank selection at rescan
When rescan is clicked, clear the selection in the bank column.
Otherwise, when the search results are returned after the
subsequent /bank_search, if the bank column was scrolled down,
the selection will be at the wrong item when the scroll state
is reset to its default state as a result of the /bank_search.