ZynBank: Don't attempt bank search at rescan
Immediately attempting a bank search when rescan is clicked
will fail, seemingly due to the response from the rescan
conflicting with the subsequent bank search. So simply refrain from
doing the search.

This means that the preset column will be empty after the rescan,
which is not really correct, but issuing a search after the rescan
doesn't help anyway, so might as well keep the code simple.