Use logmin if available
For logarithmic controls, this, aside from `min` and `max`, now also
reads the `logmin` from the Schema, if available.

The `logmin` defines the real minimum of the logarthmic scale. An
alternating `min` means "off", and is visualized by the knob making a
"jump" over the lowest 7% of the knob.

Compatibility note:
* If this commit is used with an old zyn, the logmin is simply ignored
  and the visualization is unchanged.
* If the new zyn is used with the old GUI, that old GUI will prevent
  values below "logmin" and will make the knob "jump back". This is no
  disadvantage to the old zyn, because you could not select such values
  with it, either.