Button: Add rocker mode
Rocker mode brings up two side-by-side buttons instead of one,
like a rocker or two button radio button. When one side is
active the other is not. This is intended for boolean parameters
that have two values whose counterparts are not necessarily
obvious, such as linear vs. exponential mode for the amp envs.

For these parameters, the label property needs to contain the two
values to be shown, separated by a / , e.g. "lin/log".
The left value corresponds to the 'true' value of the boolean;
the right value corresponds to the 'false' value